Move execution logic to saturn_backend crate
