+++ src-backend/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "saturn_backend"
+version = "0.1.8"
+description = "Logical bakend for the Saturn IDE."
+authors = ["desgroup"]
+license = ""
+repository = ""
+edition = "2021"
+rust-version = "1.66"
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+tokio = { version = "1.33.0", features = ["rt", "time", "macros"] }
+
+num = "0.4.1"
+byteorder = "1.5.0"
+base64 = "0.21.7"
+rand = "0.8.5"
+rand_chacha = "0.3.1"
+tokio-util = "0.7.10"
+async-trait = "0.1.77"
+
+titan = { git = "https://github.com/1whatleytay/titan.git", branch = "main" }
