+++ src-backend/src/execution.rs
@@ -1,4 +1,4 @@
-use crate::display::FlushDisplayBody;
+use crate::display::{FlushDisplayBody, read_display};
 use crate::syscall::{SyscallDelegate, SyscallResult};
 use serde::Serialize;
 use std::collections::HashSet;
@@ -13,7 +13,7 @@ use titan::execution::trackers::history::HistoryTracker;
 use titan::execution::trackers::Tracker;
 use titan::unit::instruction::InstructionDecoder;
 use titan::unit::suggestions::MemoryErrorReason;
-use crate::state::device::ExecutionState;
+use crate::device::ExecutionState;
 
 #[derive(Serialize)]
 #[serde(tag = "type")]
@@ -140,33 +140,6 @@ impl ResumeResult {
             registers: frame.registers.into(),
         }
     }
-}
-
-// NOT a tauri command.
-pub fn read_display<Mem: Memory>(
-    address: u32,
-    width: u32,
-    height: u32,
-    memory: &mut Mem,
-) -> Option<Vec<u8>> {
-    let pixels = width.checked_mul(height)?;
-
-    let mut result = vec![0u8; (pixels * 4) as usize];
-
-    for i in 0..pixels {
-        let point = address.wrapping_add(i.wrapping_mul(4));
-
-        let pixel = memory.get_u32(point).ok()?;
-
-        // Assuming little endian: 0xAARRGGBB -> [BB, GG, RR, AA] -> want [RR, GG, BB, AA]
-        let start = (i as usize) * 4;
-        result[start] = (pixel.wrapping_shr(16) & 0xFF) as u8;
-        result[start + 1] = (pixel.wrapping_shr(8) & 0xFF) as u8;
-        result[start + 2] = (pixel.wrapping_shr(0) & 0xFF) as u8;
-        result[start + 3] = 255;
-    }
-
-    Some(result)
 }
 
 pub trait ExecutionRewindable {
