+++ src-backend/src/hex_format.rs
