+++ src-backend/src/lib.rs
@@ -1,0 +1,10 @@
+pub mod execution;
+pub mod device;
+pub mod syscall;
+pub mod channels;
+pub mod display;
+pub mod keyboard;
+pub mod decode;
+pub mod build;
+pub mod hex_format;
+pub mod regions;
