+++ src-tauri/Cargo.toml
@@ -19,18 +19,19 @@ tokio = { version = "1.33.0", features = ["time", "macros"] }
 
 hex = "0.4.3"
 futures = "0.3.29"
-tokio-util = "0.7.10"
+#tokio-util = "0.7.10"
 percent-encoding = "2.3.0"
-rand = "0.8.5"
-rand_chacha = "0.3.1"
+#rand = "0.8.5"
+#rand_chacha = "0.3.1"
 sha2 = "0.10.8"
 uuid = "1.5.0"
-async-trait = "0.1.74"
-num = "0.4.1"
+#async-trait = "0.1.74"
+#num = "0.4.1"
 notify = "6.1.1"
-byteorder = "1.5.0"
+#byteorder = "1.5.0"
 base64 = "0.21.5"
 
+saturn_backend = { path = "../src-backend" }
 titan = { git = "https://github.com/1whatleytay/titan.git", branch = "main" }
 
 [features]
