+++ src-tauri/src/midi/mod.rs
@@ -4,5 +4,5 @@ mod instruments;
 mod protocol;
 
 pub use fetch::{install_instruments, midi_install, MidiProviderContainer};
-pub use handler::forward_midi;
+pub use handler::ForwardMidi;
 pub use protocol::midi_protocol;
