--- src-tauri/src/state/mod.rs
@@ -1,3 +1,0 @@
-pub mod execution;
-pub mod device;
-pub mod commands;
