+++ src-wasm/.gitignore
@@ -1,0 +1,4 @@
+# Generated by Cargo
+# will have compiled files and executables
+/target/
+
